+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Show toast in case no app is installed which can handle GPX files
 * Mi Band 4/Amazfit Bip Lite: Trim white spaces and new lines from auth key
 * Mi Band 4/Amazfit Bip Lite: Display a toast and do not try to pair if there was no auth key supplied
+* Skip service scan if supported device could be recognized without uuids during discovery
 
 #### Version 0.36.2
 * Amazfit Bip: Untested support for Lite variant